Load / Save and log file functionality implemented
Logging and query load / save implemented in this version.

Queries are saved identically to the MapInfo version of the tool.

+++ DataSelector/DataSelector/HLFileFunctions.cs
@@ -85,17 +85,41 @@ namespace HLFileFunctions
 
         public bool DeleteFile(string aFullPath)
         {
-            try
+            if (FileExists(aFullPath))
             {
-                File.Delete(aFullPath);
+                try
+                {
+                    File.Delete(aFullPath);
+                    return true;
+                }
+                catch
+                {
+                    return false;
+                }
+            }
+            else
                 return true;
-            }
-            catch
-            {
-                return false;
-            }
 
         }
 
+        public bool CreateLogFile(string aTextFile)
+        {
+            StreamWriter myWriter = new StreamWriter(aTextFile, false);
+
+            myWriter.WriteLine("Log file for Data Selector, started on " + DateTime.Now.ToString());
+            myWriter.Close();
+            myWriter.Dispose();
+            return true;
+        }
+
+        public bool WriteLine(string aTextFile, string aWriteLine)
+        {
+            StreamWriter myWriter = new StreamWriter(aTextFile, true);
+            aWriteLine = DateTime.Now.ToString() + " : " + aWriteLine;
+            myWriter.WriteLine(aWriteLine);
+            myWriter.Close();
+            myWriter.Dispose();
+            return true;
+        }
     }
 }
